prov/efa: move functions to rxr_pkt_type_base.c/h
This patch move rxr_pkt_copy_to_rx() and rxr_pkt_data_size()
from rxr_pkt_cmd.c to rxr_pkt_type_base.c.

The definition of these two functions was moved to a new
file rxr_pkt_type_base.h, along with the definitions of
3 functions whose implementations are already in rxr_pkt_type_base.c

This function alse rename rxr_pkt_copy_to_rx() to
rxr_pkt_copy_data_to_rx_entry(). The name fits the function's
role better.
