+++ prov/efa/src/rxr/rxr_pkt_type.h
@@ -127,14 +127,6 @@ static inline struct rxr_base_hdr *rxr_get_base_hdr(void *pkt)
 	return (struct rxr_base_hdr *)pkt;
 }
 
-uint32_t *rxr_pkt_connid_ptr(struct rxr_pkt_entry *pkt_entry);
-
-void rxr_pkt_init_data_from_tx_entry(struct rxr_ep *ep,
-				     struct rxr_pkt_entry *pkt_entry,
-				     size_t hdr_size,
-				     struct rxr_tx_entry *tx_entry,
-				     size_t data_offset, size_t data_size);
-
 struct rxr_ep;
 struct rdm_peer;
 struct rxr_tx_entry;
@@ -291,9 +283,12 @@ void rxr_pkt_proc_data(struct rxr_ep *ep,
 void rxr_pkt_handle_data_send_completion(struct rxr_ep *ep,
 					 struct rxr_pkt_entry *pkt_entry);
 
-
 void rxr_pkt_handle_data_recv(struct rxr_ep *ep,
 			      struct rxr_pkt_entry *pkt_entry);
+
+void rxr_pkt_handle_data_copied(struct rxr_ep *ep,
+				struct rxr_pkt_entry *pkt_entry,
+				size_t data_size);
 
 /*
  *  @brief READRSP packet header
