+++ prov/efa/src/rxr/rxr_pkt_type_base.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2020 Amazon.com, Inc. or its affiliates.
+ * Copyright (c) 2021 Amazon.com, Inc. or its affiliates.
  * All rights reserved.
  *
  * This software is available to you under a choice of one of two
@@ -31,47 +31,25 @@
  * SOFTWARE.
  */
 
-#ifndef _RXR_PKT_CMD_H
-#define _RXR_PKT_CMD_H
+#ifndef _RXR_PKT_TYPE_BASE_H
+#define _RXR_PKT_TYPE_BASE_H
 
 #include "rxr.h"
 
-ssize_t rxr_pkt_post_data(struct rxr_ep *rxr_ep, struct rxr_tx_entry *tx_entry);
+uint32_t *rxr_pkt_connid_ptr(struct rxr_pkt_entry *pkt_entry);
 
-ssize_t rxr_pkt_post_ctrl(struct rxr_ep *ep, int entry_type, void *x_entry,
-			  int ctrl_type, bool inject);
+void rxr_pkt_init_data_from_tx_entry(struct rxr_ep *ep,
+				     struct rxr_pkt_entry *pkt_entry,
+				     size_t hdr_size,
+				     struct rxr_tx_entry *tx_entry,
+				     size_t data_offset, size_t data_size);
 
-ssize_t rxr_pkt_post_ctrl_or_queue(struct rxr_ep *ep, int entry_type, void *x_entry,
-				   int ctrl_type, bool inject);
+ssize_t rxr_pkt_copy_data_to_rx_entry(struct rxr_ep *ep,
+				      struct rxr_rx_entry *rx_entry,
+				      size_t data_offset,
+				      struct rxr_pkt_entry *pkt_entry,
+				      char *data, size_t data_size);
 
-void rxr_pkt_handle_data_copied(struct rxr_ep *ep,
-				struct rxr_pkt_entry *pkt_entry,
-				size_t data_size);
-
-void rxr_pkt_handle_send_error(struct rxr_ep *ep,
-			       struct rxr_pkt_entry *pkt_entry,
-			       int err, int prov_errno);
-
-void rxr_pkt_handle_send_completion(struct rxr_ep *ep,
-				    struct rxr_pkt_entry *pkt_entry);
-
-void rxr_pkt_handle_recv_error(struct rxr_ep *ep,
-			       struct rxr_pkt_entry *pkt_entry,
-			       int err, int prov_errno);
-
-void rxr_pkt_handle_recv_completion(struct rxr_ep *ep,
-				    struct rxr_pkt_entry *pkt_entry);
-
-ssize_t rxr_pkt_wait_handshake(struct rxr_ep *ep, fi_addr_t addr, struct rdm_peer *peer);
-
-ssize_t rxr_pkt_trigger_handshake(struct rxr_ep *ep,
-				  fi_addr_t addr, struct rdm_peer *peer);
-
-#if ENABLE_DEBUG
-void rxr_pkt_print(char *prefix,
-		   struct rxr_ep *ep,
-		   struct rxr_base_hdr *hdr);
-#endif
+size_t rxr_pkt_data_size(struct rxr_pkt_entry *pkt_entry);
 
 #endif
-
